requirements.txt
haystack-ai
datasets>=2.6.1
sentence-transformers>=4.1.0
numpy<2
python-dotenv
anthropic

